requirements.txt
-e git+https://github.com/Hartz-byte/ai-job-agent.git@9f1c92385dc713923138d497182bf3f57f28a8fb#egg=ai_job_agent
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.0.0
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
diskcache==5.6.3
docxcompose==1.4.0
duckdb==1.3.2
entrypoints==0.4
Faker==37.5.3
favicon==0.7.0
filelock==3.19.1
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
htbuilder==0.9.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
llama_cpp_python==0.3.16
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
narwhals==2.1.2
numpy==2.2.6
packaging==24.2
pandas==2.3.2
pdfminer.six==20250506
pillow==10.4.0
platformdirs==4.3.8
playwright==1.54.0
plotly==6.3.0
prometheus_client==0.22.1
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.0
s3transfer==0.13.1
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.17.2
snowflake-snowpark-python==1.37.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.37.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.6
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==4.0.2
